feat(jsonConverter): Add support for converting queries with all columns selected
